+++ pkg/otlp/metrics/metrics_translator.go
@@ -26,8 +26,10 @@ import (
 	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes"
 	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes/source"
 	"github.com/DataDog/opentelemetry-mapping-go/pkg/quantile"
+	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/pmetric"
+	otelmetric "go.opentelemetry.io/otel/metric"
 	"go.uber.org/zap"
 	"golang.org/x/exp/slices"
 
@@ -38,6 +40,9 @@ import (
 const (
 	metricName             string = "metric name"
 	errNoBucketsNoSumCount string = "no buckets mode and no send count sum are incompatible"
+
+	meterName               string = "github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/metrics"
+	missingSourceMetricName string = "datadog.otlp_translator.metrics.missing_source"
 )
 
 var _ source.Provider = (*noSourceProvider)(nil)
@@ -48,11 +53,16 @@ func (*noSourceProvider) Source(context.Context) (source.Source, error) {
 	return source.Source{Kind: source.HostnameKind, Identifier: ""}, nil
 }
 
+type instruments struct {
+	missingSources otelmetric.Int64Counter
+}
+
 // Translator is a metrics translator.
 type Translator struct {
-	prevPts *ttlCache
-	logger  *zap.Logger
-	cfg     translatorConfig
+	prevPts     *ttlCache
+	logger      *zap.Logger
+	instruments instruments
+	cfg         translatorConfig
 }
 
 // Metadata specifies information about the outcome of the MapMetrics call.
@@ -62,14 +72,13 @@ type Metadata struct {
 }
 
 // NewTranslator creates a new translator with given options.
-func NewTranslator(logger *zap.Logger, options ...TranslatorOption) (*Translator, error) {
+func NewTranslator(set component.TelemetrySettings, options ...TranslatorOption) (*Translator, error) {
 	cfg := translatorConfig{
 		HistMode:                             HistogramModeDistributions,
 		SendHistogramAggregations:            false,
 		Quantiles:                            false,
 		NumberMode:                           NumberModeCumulativeToDelta,
 		InitialCumulMonoValueMode:            InitialCumulMonoValueModeAuto,
-		ResourceAttributesAsTags:             false,
 		InstrumentationLibraryMetadataAsTags: false,
 		sweepInterval:                        1800,
 		deltaTTL:                             3600,
@@ -88,10 +97,21 @@ func NewTranslator(logger *zap.Logger, options ...TranslatorOption) (*Translator
 	}
 
 	cache := newTTLCache(cfg.sweepInterval, cfg.deltaTTL)
+	meter := set.MeterProvider.Meter(meterName)
+	missingSources, err := meter.Int64Counter(
+		missingSourceMetricName,
+		otelmetric.WithDescription("OTLP metrics that are missing a source (e.g. hostname)"),
+		otelmetric.WithUnit("[metric]"),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("failed to build missing source counter: %w", err)
+	}
+
 	return &Translator{
-		prevPts: cache,
-		logger:  logger.With(zap.String("component", "metrics translator")),
-		cfg:     cfg,
+		prevPts:     cache,
+		logger:      set.Logger.With(zap.String("component", "metrics translator")),
+		instruments: instruments{missingSources: missingSources},
+		cfg:         cfg,
 	}, nil
 }
 
@@ -522,16 +542,16 @@ func (t *Translator) mapSummaryMetrics(
 	}
 }
 
-func (t *Translator) source(m pcommon.Map) (source.Source, error) {
-	src, ok := attributes.SourceFromAttrs(m)
-	if !ok {
+func (t *Translator) source(ctx context.Context, m pcommon.Map) (source.Source, bool, error) {
+	src, hasSource := attributes.SourceFromAttrs(m)
+	if !hasSource {
 		var err error
-		src, err = t.cfg.fallbackSourceProvider.Source(context.Background())
+		src, err = t.cfg.fallbackSourceProvider.Source(ctx)
 		if err != nil {
-			return source.Source{}, fmt.Errorf("failed to get fallback source: %w", err)
+			return source.Source{}, false, fmt.Errorf("failed to get fallback source: %w", err)
 		}
 	}
-	return src, nil
+	return src, hasSource, nil
 }
 
 // extractLanguageTag appends a new language tag to languageTags if a new language tag is found from the given name
@@ -643,7 +663,7 @@ func (t *Translator) MapMetrics(ctx context.Context, md pmetric.Metrics, consume
 	rms := md.ResourceMetrics()
 	for i := 0; i < rms.Len(); i++ {
 		rm := rms.At(i)
-		src, err := t.source(rm.Resource().Attributes())
+		src, hasSource, err := t.source(ctx, rm.Resource().Attributes())
 		if err != nil {
 			return metadata, err
 		}
@@ -667,6 +687,11 @@ func (t *Translator) MapMetrics(ctx context.Context, md pmetric.Metrics, consume
 		for j := 0; j < ilms.Len(); j++ {
 			ilm := ilms.At(j)
 			metricsArray := ilm.Metrics()
+
+			if !hasSource && ilm.Scope().Name() != meterName {
+				// Only count metrics if they do not come from the translator itself.
+				t.instruments.missingSources.Add(ctx, int64(metricsArray.Len()))
+			}
 
 			var additionalTags []string
 			if t.cfg.InstrumentationScopeMetadataAsTags {