+++ pkg/otlp/metrics/metrics_translator_benchmark_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 	"go.uber.org/zap"
 )
@@ -51,8 +52,10 @@ func newBenchmarkTranslator(b *testing.B, logger *zap.Logger, opts ...Translator
 		WithNumberMode(NumberModeCumulativeToDelta),
 	}, opts...)
 
+	set := componenttest.NewNopTelemetrySettings()
+	set.Logger = logger
 	tr, err := NewTranslator(
-		logger,
+		set,
 		options...,
 	)